+++ public/posts.json
@@ -9,6 +9,101 @@
       "text": "test",
       "emoji": 1,
       "date": "10:22 06/04/20"
+    },
+    {
+      "text": "ok",
+      "emoji": "angry",
+      "date": "17:15 6-4-2020"
+    },
+    {
+      "text": "ok",
+      "emoji": "angry",
+      "date": "17:15 6-4-2020"
+    },
+    {
+      "text": "hello",
+      "emoji": "angry",
+      "date": "17:16 6-4-2020"
+    },
+    {
+      "text": "guu",
+      "emoji": "angry",
+      "date": "17:16 6-4-2020"
+    },
+    {
+      "text": "hello working",
+      "date": "17:18 6-4-2020"
+    },
+    {
+      "text": "ok1",
+      "date": "17:28 6-4-2020"
+    },
+    {
+      "text": "ok2",
+      "date": "17:31 6-4-2020"
+    },
+    {
+      "text": "",
+      "date": "17:38 6-4-2020"
+    },
+    {
+      "text": "okokok",
+      "date": "17:43 6-4-2020"
+    },
+    {
+      "text": "hehehheheheh",
+      "date": "17:44 6-4-2020"
+    },
+    {
+      "text": "got it ",
+      "emoji": "happy",
+      "date": "17:47 6-4-2020"
+    },
+    {
+      "text": "ok today ",
+      "emoji": "angry",
+      "date": "9:39 7-4-2020"
+    },
+    {
+      "text": "tried the button",
+      "emoji": "sad",
+      "date": "9:41 7-4-2020"
+    },
+    {
+      "text": "ok",
+      "date": "9:47 7-4-2020"
+    },
+    {
+      "text": "",
+      "date": "9:52 7-4-2020"
+    },
+    {
+      "text": "",
+      "date": "9:55 7-4-2020"
+    },
+    {
+      "text": "ok im trying this ",
+      "emoji": "angry",
+      "date": "10:15 7-4-2020"
+    },
+    {
+      "text": "this is a second try ",
+      "emoji": "happy",
+      "date": "10:15 7-4-2020"
+    },
+    {
+      "text": "this is a third try ",
+      "emoji": "sad",
+      "date": "10:15 7-4-2020"
+    },
+    {
+      "text": "hello :p",
+      "emoji": "sad",
+      "date": "10:34 7-4-2020"
+    },
+    {
+      "text": "hello &#128543;",
+      "date": "10:34 7-4-2020"
     }
   ]
 }
