+++ public/style.css
@@ -132,7 +132,6 @@ label:hover {
 label {
   display: inline-block;
   margin-bottom: 0.5rem;
-  margin: 0 2px 0 2px;
 }
 span {
   font-size: 30px;
@@ -144,4 +143,72 @@ span {
   bottom: 0;
   width: 100%;
   height: 2.5rem; /* Footer height */
+}
+
+/* Animation music notes  */
+.music {
+  position: absolute;
+
+  margin-top: 50px;
+  animation-name: notes;
+  animation-duration: 5s;
+  animation-iteration-count: infinite;
+  font-size: 35px;
+  opacity: 0;
+  animation-direction: alternate;
+}
+#note1 {
+  left: 10%;
+  top: 60px;
+  animation-delay: 0.5s;
+  animation-timing-function: linear;
+  animation-direction: alternate-reverse;
+}
+#note2 {
+  left: 30%;
+  top: 30px;
+  animation-delay: 1s;
+  animation-timing-function: ease-out;
+  animation-direction: reverse;
+}
+#note3 {
+  left: 40%;
+  top: 90px;
+  animation-delay: 1.5s;
+  animation-timing-function: ease-in-out;
+  animation-direction: alternate;
+}
+#note4 {
+  left: 60%;
+  top: 40px;
+  animation-delay: 2s;
+  animation-timing-function: ease-in;
+  animation-direction: normal;
+}
+#note5 {
+  left: 70%;
+  top: 30px;
+  animation-delay: 2.5s;
+  animation-timing-function: ease-in;
+  animation-direction: alternate-reverse;
+}
+
+@keyframes notes {
+  0% {
+    transform: scale(1) translate(0, 0);
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+    transform: scale(1.5) translate(50%, -50%);
+  }
+  80% {
+    opacity: 0;
+    transform: scale(1) translate(100%, -100%);
+  }
+  100% {
+    transform: scale(1.5) translate(100%, -100%);
+    opacity: 0;
+    top: 80%;
+  }
 }
